# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MoscowT route-hour API",
    "version": "2.0.0"
  },
  "paths": {
...
    "/api/v1/routes": {
      "get": {
        "summary": "Get Routes",
        "operationId": "get_routes_api_v1_routes_get",
        "parameters": [
          {
            "name": "snapshotId",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Snapshotid"
            }
          },
          {
            "name": "date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Route"
                  },
                  "title": "Response Get Routes Api V1 Routes Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/exports": {
      "post": {
        "summary": "Exports",
        "operationId": "exports_api_v1_exports_post",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExportRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/jobs/{ident}": {
      "get": {
        "summary": "Get Job",
        "operationId": "get_job_api_v1_jobs__ident__get",
        "parameters": [
          {
            "name": "ident",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ident"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v2/sections": {
      "post": {
        "summary": "Sections",
        "operationId": "sections_api_v2_sections_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Scope"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v2/uploads": {
      "post": {
        "summary": "Preview",
        "operationId": "preview_api_v2_uploads_post",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UploadRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/uploads/{ident}": {
      "get": {
        "summary": "Get Upload",
        "operationId": "get_upload_api_v2_uploads__ident__get",
        "parameters": [
          {
            "name": "ident",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ident"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/uploads/{ident}/apply": {
      "post": {
        "summary": "Apply",
        "operationId": "apply_api_v2_uploads__ident__apply_post",
        "parameters": [
          {
            "name": "ident",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ident"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Idempotency-Key"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v2/datasets/{ident}": {
      "get": {
        "summary": "Dataset",
        "operationId": "dataset_api_v2_datasets__ident__get",
        "parameters": [
          {
            "name": "ident",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ident"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/datasets/{ident}/activate": {
      "post": {
        "summary": "Activate",
        "operationId": "activate_api_v2_datasets__ident__activate_post",
        "parameters": [
          {
            "name": "ident",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ident"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/routes": {
      "get": {
        "summary": "Routes",
        "operationId": "routes_api_v2_routes_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create Route",
        "operationId": "create_route_api_v2_routes_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RouteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/routes/{ident}": {
      "get": {
        "summary": "Route",
        "operationId": "route_api_v2_routes__ident__get",
        "parameters": [
          {
            "name": "ident",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ident"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v2/weather-forecasts": {
      "get": {
        "summary": "Weather Forecasts",
        "operationId": "weather_forecasts_api_v2_weather_forecasts_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      },
      "post": {
        "summary": "Fetch Weather",
        "operationId": "fetch_weather_api_v2_weather_forecasts_post",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Idempotency-Key"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/scenarios": {
      "post": {
        "summary": "Create Scenario",
        "operationId": "create_scenario_api_v2_scenarios_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScenarioSpec"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/scenarios/{ident}": {
      "get": {
        "summary": "Get Scenario",
        "operationId": "get_scenario_api_v2_scenarios__ident__get",
        "parameters": [
          {
            "name": "ident",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ident"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v2/training-runs": {
      "post": {
        "summary": "Train",
        "operationId": "train_api_v2_training_runs_post",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TrainingSpec"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/forecast-runs": {
      "post": {
        "summary": "Forecast",
        "operationId": "forecast_api_v2_forecast_runs_post",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ForecastSpec"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "Forecast Runs",
        "operationId": "forecast_runs_api_v2_forecast_runs_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/api/v2/forecast-runs/{ident}": {
      "get": {
        "summary": "Forecast Run",
        "operationId": "forecast_run_api_v2_forecast_runs__ident__get",
        "parameters": [
          {
            "name": "ident",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ident"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/forecast-runs/{ident}/evaluation": {
      "get": {
        "summary": "Evaluate",
        "operationId": "evaluate_api_v2_forecast_runs__ident__evaluation_get",
        "parameters": [
          {
            "name": "ident",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ident"
            }
          },
          {
            "name": "truth_dataset_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Truth Dataset Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/forecasts/query": {
      "post": {
        "summary": "Query",
        "operationId": "query_api_v2_forecasts_query_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PeriodExportSpec"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Route": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "number": {
            "type": "string",
            "title": "Number"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "isTarget": {
            "type": "boolean",
            "title": "Istarget"
          },
          "hasData": {
            "type": "boolean",
            "title": "Hasdata"
          },
          "hasGeometry": {
            "type": "boolean",
            "title": "Hasgeometry"
          },
          "historySupport": {
            "type": "string",
            "title": "Historysupport"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "id",
          "number",
          "name",
          "isTarget",
          "hasData",
          "hasGeometry",
          "historySupport"
        ],
        "title": "Route"
      },
...
      "UploadRequest": {
        "properties": {
          "blob_id": {
            "type": "string",
            "title": "Blob Id"
          },
          "spec": {
            "$ref": "#/components/schemas/ImportSpec"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "blob_id",
          "spec"
        ],
        "title": "UploadRequest"
      },
...
    }
  }
}